# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: GeoTravel API
  version: "0.1.0"
  description: >
    Contrato REST del backend. Actualizar este archivo en el mismo PR que agregue
    o cambie un endpoint (ver CONTRIBUTING.md). Base path real: /geotravel/api
    (definido en backend/src/main/webapp/WEB-INF/web.xml).
servers:
  - url: http://localhost:8080/geotravel/api

paths:
  /health:
    get:
      summary: Estado del servicio
      responses:
        "200":
          description: OK

  /zonas:
    get:
      summary: Lista de zonas turísticas (geom en GeoJSON lon/lat, EPSG:4326)
      tags: [zonas]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Zona" }
    post:
      summary: Crear zona (valida datos y que no se superponga con otra)
      tags: [zonas]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ZonaInput" }
      responses:
        "201": { description: Creada, content: { application/json: { schema: { $ref: "#/components/schemas/IdCreado" } } } }
        "400": { $ref: "#/components/responses/DatosInvalidos" }
        "409": { $ref: "#/components/responses/ZonaSuperpuesta" }

  /zonas/{id}:
    put:
      summary: Editar zona (misma validación; la zona no cuenta como superpuesta consigo misma)
      tags: [zonas]
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ZonaInput" }
      responses:
        "200": { description: Actualizada }
        "400": { $ref: "#/components/responses/DatosInvalidos" }
        "404": { $ref: "#/components/responses/NoExiste" }
        "409": { $ref: "#/components/responses/ZonaSuperpuesta" }
    delete:
      summary: Eliminar zona
      tags: [zonas]
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses:
        "204": { description: Eliminada }
        "404": { $ref: "#/components/responses/NoExiste" }

  /atracciones:
    get:
      summary: Lista de atracciones turísticas (geom en GeoJSON lon/lat, EPSG:4326)
      tags: [atracciones]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Atraccion" }
    post:
      summary: Crear atracción
      tags: [atracciones]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AtraccionInput" }
      responses:
        "201": { description: Creada, content: { application/json: { schema: { $ref: "#/components/schemas/IdCreado" } } } }
        "400": { $ref: "#/components/responses/DatosInvalidos" }

  /atracciones/{id}:
    put:
      summary: Editar atracción
      tags: [atracciones]
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AtraccionInput" }
      responses:
        "200": { description: Actualizada }
        "400": { $ref: "#/components/responses/DatosInvalidos" }
        "404": { $ref: "#/components/responses/NoExiste" }
    delete:
      summary: Eliminar atracción (se quita también de los itinerarios de recorridos)
      tags: [atracciones]
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses:
        "204": { description: Eliminada }
        "404": { $ref: "#/components/responses/NoExiste" }

  /fotos:
    post:
      summary: Subir la foto de una atracción. Devuelve una URL para usar en fotoUrl (POST/PUT /atracciones).
      tags: [atracciones]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [archivo]
              properties:
                archivo: { type: string, format: binary, description: "JPG, PNG, GIF o WEBP, máx. 5 MB. El tipo se valida por el contenido." }
      responses:
        "201":
          description: Subida
          content:
            application/json:
              schema:
                type: object
                properties:
                  url: { type: string, example: "http://localhost:8080/geotravel/api/fotos/3f2a...c1.jpg" }
        "400": { $ref: "#/components/responses/DatosInvalidos" }
        "413": { description: La imagen supera 5 MB }

  /fotos/{nombre}:
    get:
      summary: Descargar una foto subida
      tags: [atracciones]
      parameters: [{ name: nombre, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: La imagen }
        "404": { $ref: "#/components/responses/NoExiste" }

  /recorridos:
    get:
      summary: Lista de recorridos, con estado efectivo (estacionalidad resuelta)
      tags: [recorridos]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Recorrido" }
    post:
      summary: "TODO (P4): crear recorrido + itinerario de atracciones"
      tags: [recorridos]
      responses:
        "201": { description: Creado }

  /recorridos/{id}/avanzar:
    post:
      summary: Avanza el recorrido al siguiente estado válido (Pendiente -> Disponible -> Cancelado)
      tags: [recorridos]
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses:
        "200": { description: Estado actualizado }
        "404": { description: No existe }
        "409": { description: No hay transición válida desde el estado actual }

  /recorridos/{id}/historial:
    get:
      summary: "TODO (P4): histórico de estados del recorrido"
      tags: [recorridos]
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses:
        "200": { description: OK }

  /consultas/recorridos-por-zona/{zonaId}:
    get:
      summary: Recorridos con al menos una atracción dentro de la zona
      tags: [consultas]
      parameters: [{ name: zonaId, in: path, required: true, schema: { type: integer } }]
      responses:
        "200": { description: OK }

  /consultas/zonas-mas-activas:
    get:
      summary: Zonas ordenadas por cantidad de recorridos "Disponible"
      tags: [consultas]
      responses:
        "200": { description: OK }

  /consultas/puntos-populares:
    get:
      summary: Atracciones ordenadas por cantidad de recorridos que las incluyen
      tags: [consultas]
      responses:
        "200": { description: OK }

  /consultas/recorrido-mas-cercano:
    get:
      summary: Recorrido cuya atracción más cercana está al punto (x, y) dado
      tags: [consultas]
      parameters:
        - { name: x, in: query, required: true, schema: { type: number }, description: "Este, EPSG:32721" }
        - { name: y, in: query, required: true, schema: { type: number }, description: "Norte, EPSG:32721" }
      responses:
        "200": { description: OK }

  /consultas/zona-por-punto:
    get:
      summary: Zona que contiene el punto (x, y) dado -- resolver antes la dirección a coordenada
      tags: [consultas]
      parameters:
        - { name: x, in: query, required: true, schema: { type: number } }
        - { name: y, in: query, required: true, schema: { type: number } }
      responses:
        "200": { description: OK }

  /consultas/reportes/recorridos-por-zona:
    get:
      summary: "TODO (P5): reporte, filtrable por estado, agrupado por zona"
      tags: [consultas]
      parameters:
        - { name: estado, in: query, required: false, schema: { type: string } }
      responses:
        "200": { description: OK }

components:
  responses:
    DatosInvalidos:
      description: Datos inválidos (nombre vacío, nivel fuera de 1-5, clasificación desconocida, geometría o URL inválida)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    ZonaSuperpuesta:
      description: 'La zona se superpone con otra existente; el mensaje indica con cuál (ej. La zona se superpone con "Pocitos")'
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NoExiste:
      description: No existe
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
    IdCreado:
      type: object
      properties:
        id: { type: integer }
    ZonaInput:
      type: object
      required: [nombre, nivelAtractivo, geomGeoJson]
      properties:
        nombre: { type: string }
        descripcion: { type: string }
        nivelAtractivo: { type: integer, minimum: 1, maximum: 5, description: "1 = mayor atractivo" }
        observaciones: { type: string }
        geomGeoJson: { type: string, description: 'GeoJSON Polygon en lon/lat (EPSG:4326) como TEXTO (JSON.stringify)' }
    AtraccionInput:
      type: object
      required: [nombre, clasificacion, geomGeoJson]
      properties:
        nombre: { type: string }
        descripcion: { type: string }
        clasificacion: { type: string, enum: [cultural, gastronomica, natural, historica] }
        fotoUrl: { type: string, description: "Opcional. URL http(s), p. ej. la devuelta por POST /fotos" }
        geomGeoJson: { type: string, description: 'GeoJSON Point en lon/lat (EPSG:4326) como TEXTO (JSON.stringify)' }
    Zona:
      type: object
      properties:
        id: { type: integer }
        nombre: { type: string }
        descripcion: { type: string }
        nivelAtractivo: { type: integer, minimum: 1, maximum: 5 }
        observaciones: { type: string }
        geom: { type: string, description: "GeoJSON Polygon en lon/lat (EPSG:4326), como texto: hacer JSON.parse" }
    Atraccion:
      type: object
      properties:
        id: { type: integer }
        nombre: { type: string }
        descripcion: { type: string }
        clasificacion: { type: string }
        fotoUrl: { type: string }
        geom: { type: string, description: "GeoJSON Point en lon/lat (EPSG:4326), como texto: hacer JSON.parse" }
    Recorrido:
      type: object
      properties:
        id: { type: integer }
        nombre: { type: string }
        duracionMin: { type: integer }
        guiaResponsable: { type: string }
        tipoExperiencia: { type: string }
        mesInicio: { type: integer }
        mesFin: { type: integer }
        estado:
          type: string
          enum: [Pendiente, Disponible, "Fuera de estacion", Cancelado]
        estadoEfectivo:
          type: string
          description: estado ya resuelto contra la estacionalidad actual